==> Lessons/Lessons_NK edited/09_DataVisualization_NK.Rmd <==
---
title: "9: Data Visualization"
author: "Environmental Data Analytics | Kateri Salk"
date: "Spring 2019"
output: pdf_document
geometry: margin=2.54cm
editor_options: 
  chunk_output_type: console
---

## LESSON OBJECTIVES
1. Perform simple data visualizations in the R package `ggplot`
2. Develop skills to adjust aesthetics and layers in graphs
3. Apply a decision tree framework for appropriate graphing methods

## SET UP YOUR DATA ANALYSIS SESSION

```{r, message = FALSE, warning = FALSE}
getwd()
library(tidyverse) #ggplot is included in the tidyverse package

PeterPaul.chem.nutrients <- read.csv("./Data/Processed/NTL-LTER_Lake_Chemistry_Nutrients_PeterPaul_Processed.csv")
PeterPaul.nutrients.gathered <- read.csv("./Data/Processed/NTL-LTER_Lake_Nutrients_PeterPaulGathered_Processed.csv")
PeterPaul.chem.nutrients.summaries <- read.csv("./Data/Processed/NTL-LTER_Lake_Summaries_PeterPaul_Processed.csv")
EPAair <- read.csv("./Data/Processed/EPAair_O3_PM25_NC1718_Processed.csv")

EPAair$Date <- as.Date(EPAair$Date, format = "%Y-%m-%d")
PeterPaul.chem.nutrients$sampledate <- as.Date(PeterPaul.chem.nutrients$sampledate, format = "%Y-%m-%d")

```

## GGPLOT

ggplot, called from the package `ggplot2`, is a graphing and image generation tool in R. This package is part of tidyverse. While base R has graphing capabilities, ggplot has the capacity for a wider range and more sophisticated options for graphing. 
ggplot has only a few rules: 

* The first line of ggplot code always starts with `ggplot()`
* A data frame must be specified within the `ggplot()` function. Additional datasets can be specified in subsequent layers.
* Aesthetics must be specified, most commonly x and y variables but including others. Aesthetics can be specified in the `ggplot()` function or in subsequent layers.
* Additional layers must be specified to fill the plot. 

### Geoms

Here are some commonly used layers for plotting in ggplot: 

* geom_bar
* geom_histogram
* geom_freqpoly
* geom_boxplot
* geom_violin
* geom_dotplot
* geom_point
* geom_errorbar
* geom_smooth
* geom_line
* geom_area
* geom_abline (plus geom_hline and geom_vline)
* geom_text

### Aesthetics
Here are some commonly used aesthetic types that can be manipulated in ggplot: 

* color
* fill
* shape
* size
* transparency

### Plotting continuous variables over time: Scatterplot
```{r, fig.height = 3, fig.width = 4, warning = FALSE}
# Scatterplot
ggplot(EPAair, aes(x = Date, y = Ozone_Daily_AQI)) + 
  geom_point()

O3plot <- ggplot(EPAair) +
  geom_point(aes(x = Date, y = Ozone_Daily_AQI))
print(O3plot) #We indicate print because we want it to show up in our pdf doc

# Fix this code
O3plot2 <- ggplot(EPAair) +
  geom_point(aes(x = Date, y = Ozone_Daily_AQI) , color = "Blue") # if we are not defining colour based on data within our data frame than we can define it ouside of the aes function
print(O3plot2)

# Add additional variables
PMplot <- 
  ggplot(EPAair, aes(x = as.factor(Month), y = PM2.5_Daily_AQI, shape = as.factor(Year), color = Site.Name)) +
  geom_point() #month and year are changed to factor because they are integers in the dataset
print(PMplot)

# Separate plot with facets
PMplot.faceted <-
  ggplot(EPAair, aes(x = as.factor(Month), y = PM2.5_Daily_AQI, shape = as.factor(Year))) +
  geom_point() +
  facet_wrap(vars(Site.Name), nrow = 3) #default is ploting into columns
print(PMplot.faceted)

PMplot.faceted2 <-
  ggplot(EPAair, aes(x = as.factor(Month), y = PM2.5_Daily_AQI)) +
  geom_point() +
  facet_grid(Site.Name ~ Year) #order of site.name and year matters
print(PMplot.faceted2)

# Filter dataset within plot building
O3plot.Blackstone <- 
  ggplot(subset(EPAair, Site.Name == "Blackstone"), aes(x = Date, y = Ozone_Daily_AQI)) + 
  geom_point() +
  geom_line() #adding a line layer to connect points
print(O3plot.Blackstone)

# Exercise: build your own scatterplots of PeterPaul.chem.nutrients

# 1. 
# Plot surface temperatures by day of  year. 
# Color your points by year, and facet by lake in two rows.
Pet.Paul.chem.nut.plot1 <- ggplot(PeterPaul.chem.nutrients, aes(x=daynum, y=temperature_C, color=year4))+
  geom_point() +
  facet_wrap(vars(lakename), nrow = 2)
print(Pet.Paul.chem.nut.plot1)

#2. 
# Plot temperature by date. Color your points by depth.
# Change the size of your point to 0.5
Pet.Paul.chem.nut.plot2 <- ggplot(PeterPaul.chem.nutrients, aes(x=sampledate, y=temperature_C, color=depth))+
  geom_point(size = 0.5)
print(Pet.Paul.chem.nut.plot2)

```
### Plotting the relationship between two continuous variables: Scatterplot

```{r, fig.height = 3, fig.width = 4, warning = FALSE}
# Scatterplot
lightvsDO <- 
  ggplot(PeterPaul.chem.nutrients, aes(x = irradianceWater, y = dissolvedOxygen)) +
  geom_point()
print(lightvsDO)

#zooming in to a certain region of the plot
# Adjust axes
lightvsDOfixed <- 
  ggplot(PeterPaul.chem.nutrients, aes(x = irradianceWater, y = dissolvedOxygen)) +
  geom_point() +
  xlim(0, 250) +
  ylim(0, 20)
print(lightvsDOfixed)

# Depth in the fields of limnology and oceanography is on a reverse scale
tempvsdepth <- 
  #ggplot(PeterPaul.chem.nutrients, aes(x = temperature_C, y = depth)) +
  ggplot(PeterPaul.chem.nutrients, aes(x = temperature_C, y = depth, color = daynum)) +
  geom_point() +
  scale_y_reverse() #reverses the y axis scale
print(tempvsdepth)

NvsP <-
  ggplot(PeterPaul.chem.nutrients, aes(x = tp_ug, y = tn_ug, color = depth)) +
  geom_point() +
  #its important to specify that it is a linear line that we want (method =lm)
  geom_smooth(method = lm) +
  geom_abline(aes(slope = 16, intercept = 0), lty = 2) #adding a dotted line line to the plot 
print(NvsP)

# Exercise: Plot relationships between air quality measurements

# 1. 
# Plot AQI values for ozone by PM2.5, colored by site. 
# Add a line of best fit for the linear regression of these variables.
AQI.plot <- 
  ggplot(EPAair, aes(x=PM2.5_Daily_AQI, y=Ozone_Daily_AQI, color = Site.Name)) +
   geom_point() +
  geom_smooth(method = lm)
print(AQI.plot)
```

### Plotting continuous vs. categorical variables

```{r, fig.height = 3, fig.width = 4, warning = FALSE}
 # Barplot + error bars
PeterPaul.nutrient.summaries <- PeterPaul.nutrients.gathered %>%
  group_by(lakename, nutrient) %>%
  summarise(sd = sd(concentration), 
            mean = mean(concentration))

Nutrientplot <- 
  ggplot(PeterPaul.nutrients.gathered) +
  geom_bar(aes(x = lakename, y = concentration, fill = as.factor(nutrient)), # why did we use fill? - so that the inside of the bar chart would be colour coded and not just the outline
           position = "dodge", stat = "summary", fun.y = "mean")             # what's happening here? - stat- summaposition dodge ensures the bars are separated out and not stacked on each other. fun.y mean esures the value we are ploting is the mean of the data
print(Nutrientplot)

#a bar chart requires a function defining the stat it is using,
#stat =summary requires plot to compute the summary statistics for the variables we want to plot
#stats=identity is used when the variable has one value and you can point the plot to that one valuemakes. In the example below this is the mean computed in the summaries dataset. 

Nutrientplot2 <- 
  ggplot(PeterPaul.nutrient.summaries, aes(x = lakename, y = mean, fill = as.factor(nutrient))) + #
  geom_bar(stat = "identity", position = "dodge") + # what does the stat command do?
  geom_errorbar(aes(ymin = mean-sd, ymax = mean+sd), # how do we specify error bars? - we specify the minimum of the error bar and the maximum to show standar deviation range
                 position = "dodge")
print(Nutrientplot2)

# Are there more effective ways to produce summary stats for categories?
#many things in data analytics are moving away from bars with error bars because you can more efficiently display summary stats
#Better alternatives are below
# Box and whiskers plot
Nutrientplot3 <-
  ggplot(PeterPaul.nutrients.gathered, aes(x = lakename, y = concentration)) +
  geom_boxplot(aes(color = nutrient)) # Why didn't we use "fill"? fill colors the inside of the box but it leaves the outline as black
print(Nutrientplot3)

# Dot plot
Nutrientplot4 <-
  ggplot(PeterPaul.nutrients.gathered, aes(x = lakename, y = concentration)) +
  geom_dotplot(aes(color = nutrient), binaxis = "y", binwidth = 1, 
               stackdir = "center", position = "dodge") #wide spots of the dot plot show points with several points. useful for showing density of particular points
print(Nutrientplot4)

# Violin plot
Nutrientplot5 <-
  ggplot(PeterPaul.nutrients.gathered, aes(x = lakename, y = concentration)) +
  geom_violin(aes(color = nutrient)) #similar to a dot plot but is a line that represents the distribution of the points
print(Nutrientplot5)

# Frequency polygons
# Using a tidy dataset
Nutrientplot6 <-
  ggplot(PeterPaul.chem.nutrients) +
  geom_freqpoly(aes(x = tn_ug), color = "black") +
  geom_freqpoly(aes(x = tp_ug), color = "darkblue") +
  geom_freqpoly(aes(x = nh34), color = "darkgray") +
  geom_freqpoly(aes(x = no23), color = "gray") +
  geom_freqpoly(aes(x = po4), color = "blue") 
print(Nutrientplot6) #it is important to specify the color because default is black and all lines will be ploted with black

# Using a gathered dataset
Nutrientplot7 <-   
  ggplot(PeterPaul.nutrients.gathered) +
  geom_freqpoly(aes(x = concentration, color = nutrient))
print(Nutrientplot7)

# Exercise: Plot distributions of AQI values for EPAair

# 1. 
# Create a bar chart plus standard deviation error bars for PM2.5, divided by year. 
# Create separate bars for each site. 
EPAairsummary <- EPAair %>%
  group_by(Site.Name, Year) %>%
  summarise(sd = (sd(PM2.5_Daily_AQI, na.rm = TRUE)),
            mean = mean(PM2.5_Daily_AQI, na.rm = TRUE))

plot1 <- 
  ggplot(EPAairsummary, aes(x = as.factor(Year), y =mean, fill = as.factor(Site.Name))) + #
  geom_bar(stat = "identity", position = "dodge") + # what does the stat command do?
  geom_errorbar(aes(ymin = mean-sd, ymax = mean+sd), # how do we specify error bars? - we specify the minimum of the error bar and the maximum to show standar deviation range
                 position = "dodge")
print(plot1)
# 2. 
# Create a new plot that better depicts the distribution of PM2.5 concentrations. 
# Divide your graph by year and site.

```
